add test coverage for OTEL logging
